Updated six more DynamoDB Examples.
