--- dotnetv3/dynamodb/FromSQL/AddItems/app.config
@@ -1,11 +1,0 @@
-<configuration>
- <appSettings>
-    <add key="Region" value="us-west-2" />
-    <add key="Table" value="CustomersOrdersProducts" />
-   <add key="Index" value="100" />
-   <add key="Filename" value="../../../Customers.csv" />
-    <add key="Customers" value="../../../Customers.csv" />
-    <add key="Orders" value="../../../Orders.csv" />
-    <add key="Products" value="../../../Products.csv" />
- </appSettings>
-</configuration>
