+++ dotnetv3/dynamodb/FromSQL/AddItemsExample/AddItemsExample.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31624.102
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AddItemsExample", "AddItemsExample\AddItemsExample.csproj", "{E40E0DA2-D9E2-48AA-94CA-E5E43C931902}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E40E0DA2-D9E2-48AA-94CA-E5E43C931902}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E40E0DA2-D9E2-48AA-94CA-E5E43C931902}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E40E0DA2-D9E2-48AA-94CA-E5E43C931902}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E40E0DA2-D9E2-48AA-94CA-E5E43C931902}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {85607BA1-12AE-4D9F-B14C-62CBFB39E4DC}
+	EndGlobalSection
+EndGlobal
