+++ dotnetv3/dynamodb/FromSQL/AddItemsExample/customers.csv
