+++ dotnetv3/dynamodb/FromSQL/AddItemsExample/products.csv
