+++ dotnetv3/dynamodb/FromSQL/CreateTableExample/CreateTableExample.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31624.102
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CreateTableExample", "CreateTableExample\CreateTableExample.csproj", "{1257DC5F-134A-4E74-83F0-51319F4F8F51}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{1257DC5F-134A-4E74-83F0-51319F4F8F51}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1257DC5F-134A-4E74-83F0-51319F4F8F51}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1257DC5F-134A-4E74-83F0-51319F4F8F51}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1257DC5F-134A-4E74-83F0-51319F4F8F51}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {01A4E395-7E1B-4F2E-900B-422EC8837D6C}
+	EndGlobalSection
+EndGlobal
