+++ dotnetv3/dynamodb/FromSQL/CreateTableExample/app.config
