--- dotnetv3/dynamodb/FromSQL/CreateTableTest/CreateTableTest.cs
@@ -1,58 +1,0 @@
-﻿// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved. 
-// SPDX - License - Identifier: Apache - 2.0
-using System.Net;
-using System.Threading;
-using System.Threading.Tasks;
-
-using Amazon.DynamoDBv2;
-using Amazon.DynamoDBv2.Model;
-
-using Moq;
-
-using Xunit;
-using Xunit.Abstractions;
-
-namespace CreateTableTest
-{        
-    public class CreateTableTest
-    {
-        private readonly ITestOutputHelper _output;
-
-        public CreateTableTest(ITestOutputHelper output)
-        {
-            this._output = output;
-        }
-
-        readonly string _tableName = "testtable";
-
-        private IAmazonDynamoDB CreateMockDynamoDbClient()
-        {
-            var mockDynamoDbClient = new Mock<IAmazonDynamoDB>();
-
-            mockDynamoDbClient.Setup(client => client.CreateTableAsync(
-                It.IsAny<CreateTableRequest>(),
-                It.IsAny<CancellationToken>()))
-                .Callback<CreateTableRequest, CancellationToken>((request, token) =>
-                {})
-                .Returns((CreateTableRequest r, CancellationToken token) =>
-                {
-                    return Task.FromResult(new CreateTableResponse { HttpStatusCode = HttpStatusCode.OK });
-                });
-
-            return mockDynamoDbClient.Object;
-        }
-
-        [Fact]
-        public async Task CheckCreateTable()
-        {
-            IAmazonDynamoDB client = CreateMockDynamoDbClient();
-
-            var result = await CreateTable.CreateTable.MakeTableAsync(client, _tableName);
-
-            bool ok = result.HttpStatusCode == HttpStatusCode.OK;
-            Assert.True(ok, "Could NOT create table " + _tableName);
-
-            _output.WriteLine("Created table");
-        }
-    }
-}
