+++ dotnetv3/dynamodb/FromSQL/DeleteItemsExample/DeleteItemsExample.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31624.102
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeleteItemsExample", "DeleteItemsExample\DeleteItemsExample.csproj", "{266B0DF7-7CD7-4252-B5F6-EB500A6B5A75}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{266B0DF7-7CD7-4252-B5F6-EB500A6B5A75}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{266B0DF7-7CD7-4252-B5F6-EB500A6B5A75}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{266B0DF7-7CD7-4252-B5F6-EB500A6B5A75}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{266B0DF7-7CD7-4252-B5F6-EB500A6B5A75}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {64B8B910-2D71-499F-9F6E-1D93F98ED9D3}
+	EndGlobalSection
+EndGlobal
