+++ dotnetv3/dynamodb/FromSQL/DeleteTableExample/DeleteTableExample.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31624.102
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DeleteTableExample", "DeleteTableExample\DeleteTableExample.csproj", "{7C005BB4-EA02-4F6D-851F-29A62590AF61}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{7C005BB4-EA02-4F6D-851F-29A62590AF61}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7C005BB4-EA02-4F6D-851F-29A62590AF61}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7C005BB4-EA02-4F6D-851F-29A62590AF61}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7C005BB4-EA02-4F6D-851F-29A62590AF61}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {CC262DED-CAD4-4392-83DD-EC6D37D749E4}
+	EndGlobalSection
+EndGlobal
