--- dotnetv3/dynamodb/FromSQL/ListItems/app.config
@@ -1,6 +1,0 @@
-<configuration>
- <appSettings>
-    <add key="Region" value="us-west-2" />
-    <add key="Table" value="CustomersOrdersProducts" />
- </appSettings>
-</configuration>
