+++ dotnetv3/dynamodb/FromSQL/ListItemsExample/ListItemsExample.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31624.102
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ListItemsExample", "ListItemsExample\ListItemsExample.csproj", "{CF9EBFF8-566C-45FE-935E-9E2C4AFB87CF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CF9EBFF8-566C-45FE-935E-9E2C4AFB87CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CF9EBFF8-566C-45FE-935E-9E2C4AFB87CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CF9EBFF8-566C-45FE-935E-9E2C4AFB87CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CF9EBFF8-566C-45FE-935E-9E2C4AFB87CF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E3006B21-0FF7-44BE-8FE9-BCB501BB097F}
+	EndGlobalSection
+EndGlobal
