+++ dotnetv3/dynamodb/FromSQL/ListItemsExample/app.config
