--- dotnetv3/dynamodb/FromSQL/UpdateItemDataModel/app.config
@@ -1,5 +1,0 @@
-<configuration>
- <appSettings>
-    <add key="Region" value="us-west-2" />
- </appSettings>
-</configuration>
