+++ dotnetv3/dynamodb/FromSQL/UpdateItemDataModelExample/UpdateItemDataModelExample.sln
@@ -1,0 +1,34 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "UpdateItemDataModelExample", "UpdateItemDataModelExample\UpdateItemDataModelExample.csproj", "{9B412054-C052-44E1-A840-FC177B327981}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{9B412054-C052-44E1-A840-FC177B327981}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{9B412054-C052-44E1-A840-FC177B327981}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{9B412054-C052-44E1-A840-FC177B327981}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{9B412054-C052-44E1-A840-FC177B327981}.Debug|x64.Build.0 = Debug|Any CPU
+		{9B412054-C052-44E1-A840-FC177B327981}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{9B412054-C052-44E1-A840-FC177B327981}.Debug|x86.Build.0 = Debug|Any CPU
+		{9B412054-C052-44E1-A840-FC177B327981}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{9B412054-C052-44E1-A840-FC177B327981}.Release|Any CPU.Build.0 = Release|Any CPU
+		{9B412054-C052-44E1-A840-FC177B327981}.Release|x64.ActiveCfg = Release|Any CPU
+		{9B412054-C052-44E1-A840-FC177B327981}.Release|x64.Build.0 = Release|Any CPU
+		{9B412054-C052-44E1-A840-FC177B327981}.Release|x86.ActiveCfg = Release|Any CPU
+		{9B412054-C052-44E1-A840-FC177B327981}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
